LPS-74657 Add new class CheckStyleLogger
