+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checkstyle/util/CheckStyleLogger.java
@@ -1,0 +1,90 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.source.formatter.checkstyle.util;
+
+import com.liferay.portal.kernel.util.CharPool;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.source.formatter.SourceFormatterMessage;
+
+import com.puppycrawl.tools.checkstyle.DefaultLogger;
+import com.puppycrawl.tools.checkstyle.api.AuditEvent;
+
+import java.io.OutputStream;
+
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+import java.util.Set;
+import java.util.TreeSet;
+
+/**
+ * @author Hugo Huijser
+ */
+public class CheckStyleLogger extends DefaultLogger {
+
+	public CheckStyleLogger(
+		OutputStream outputStream, boolean closeStreamsAfterUse,
+		String baseDirName) {
+
+		super(outputStream, closeStreamsAfterUse);
+
+		_baseDirName = baseDirName;
+	}
+
+	@Override
+	public void addError(AuditEvent auditEvent) {
+		_sourceFormatterMessages.add(
+			new SourceFormatterMessage(
+				_getRelativizedFileName(auditEvent), auditEvent.getMessage(),
+				auditEvent.getLine()));
+
+		super.addError(auditEvent);
+	}
+
+	public Set<SourceFormatterMessage> getSourceFormatterMessages() {
+		return _sourceFormatterMessages;
+	}
+
+	private Path _getAbsoluteNormalizedPath(String pathName) {
+		Path path = Paths.get(pathName);
+
+		path = path.toAbsolutePath();
+
+		return path.normalize();
+	}
+
+	private String _getRelativizedFileName(AuditEvent auditEvent) {
+		if (Validator.isNull(_baseDirName)) {
+			return auditEvent.getFileName();
+		}
+
+		Path baseDirPath = _getAbsoluteNormalizedPath(_baseDirName);
+
+		Path relativizedPath = baseDirPath.relativize(
+			_getAbsoluteNormalizedPath(auditEvent.getFileName()));
+
+		return _baseDirName +
+			StringUtil.replace(
+				relativizedPath.toString(), CharPool.BACK_SLASH,
+				CharPool.SLASH);
+	}
+
+	private static final Set<SourceFormatterMessage> _sourceFormatterMessages =
+		new TreeSet<>();
+
+	private final String _baseDirName;
+
+}
